# specs/001-it-inventory/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: IT Inventory Management System API
  description: RESTful API for managing IT equipment inventory, assignments, and reporting
  version: 1.0.0
  contact:
    name: IT Department
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://inventory.company.com/api
    description: Production server
...
paths:
...
  /assignments:
    get:
      tags: [Assignments]
      summary: List assignments
      operationId: listAssignments
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter active (not returned) assignments
        - name: itemId
          in: query
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssignmentDetail'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Assignments]
      summary: Create assignment (assign item to employee)
      operationId: createAssignment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDetail'
        '400':
          description: Item not available or already assigned
        '403':
          $ref: '#/components/responses/Forbidden'
  /assignments/{id}/return:
    put:
      tags: [Assignments]
      summary: Process item return
      operationId: returnItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnItemRequest'
      responses:
        '200':
          description: Return processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDetail'
        '400':
          description: Assignment already returned
        '403':
          $ref: '#/components/responses/Forbidden'
  /assignments/{id}/transfer:
    put:
      tags: [Assignments]
      summary: Transfer item to another employee
      operationId: transferItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferItemRequest'
      responses:
        '200':
          description: Transfer completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  previousAssignment:
                    $ref: '#/components/schemas/AssignmentDetail'
                  newAssignment:
                    $ref: '#/components/schemas/AssignmentDetail'
        '400':
          description: Assignment already returned
        '403':
          $ref: '#/components/responses/Forbidden'
  /assignments/{id}/acknowledge:
    post:
      tags: [Assignments]
      summary: Acknowledge assignment receipt
      operationId: acknowledgeAssignment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: Acknowledgment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDetail'

  # ==================== REPORTS ====================
...
  /reports/assignments:
    get:
      tags: [Reports]
      summary: Assignment report with date range
      operationId: getAssignmentReport
      security:
        - bearerAuth: []
      parameters:
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Assignment report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentReport'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AssignmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ItemCategory:
      type: string
      enum: [LAPTOP, MONITOR, KEYBOARD, MOUSE, HEADSET, PHONE, TABLET, CABLES, SOFTWARE_LICENSE, OTHER]
    ItemStatus:
      type: string
      enum: [AVAILABLE, ASSIGNED, RESERVED, IN_REPAIR, DECOMMISSIONED, LOST_STOLEN]
    ItemCondition:
      type: string
      enum: [NEW, GOOD, FAIR, NEEDS_REPAIR, DECOMMISSIONED]

    # ========== USER ==========
...
    ItemSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetId:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/ItemCategory'
        status:
          $ref: '#/components/schemas/ItemStatus'
        condition:
          $ref: '#/components/schemas/ItemCondition'
        brand:
          type: string
        model:
          type: string
        thumbnailUrl:
          type: string
        currentAssignee:
          type: string
          description: Employee name if assigned
...
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        title:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        assignmentCount:
          type: integer
          description: Number of currently assigned items
...
    AssignmentDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        item:
          $ref: '#/components/schemas/ItemSummary'
        employee:
          $ref: '#/components/schemas/Employee'
        assignedBy:
          type: string
          description: User name
        assignedAt:
          type: string
          format: date-time
        expectedReturnAt:
          type: string
          format: date-time
        purpose:
          type: string
        conditionAtAssignment:
          $ref: '#/components/schemas/ItemCondition'
        acknowledged:
          type: boolean
        acknowledgedAt:
          type: string
          format: date-time
        notes:
          type: string
        returnedAt:
          type: string
          format: date-time
        returnedBy:
          type: string
          description: User name
        conditionAtReturn:
          $ref: '#/components/schemas/ItemCondition'
        returnNotes:
          type: string
...
    CreateAssignmentRequest:
      type: object
      required: [itemId, employeeId, conditionAtAssignment]
      properties:
        itemId:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        assignedAt:
          type: string
          format: date-time
          description: Defaults to now
        expectedReturnAt:
          type: string
          format: date-time
        purpose:
          type: string
        conditionAtAssignment:
          $ref: '#/components/schemas/ItemCondition'
        notes:
          type: string
    ReturnItemRequest:
      type: object
      required: [conditionAtReturn]
      properties:
        returnedAt:
          type: string
          format: date-time
          description: Defaults to now
        conditionAtReturn:
          $ref: '#/components/schemas/ItemCondition'
        returnNotes:
          type: string
    TransferItemRequest:
      type: object
      required: [toEmployeeId, reason]
      properties:
        toEmployeeId:
          type: string
          format: uuid
        reason:
          type: string
        transferredAt:
          type: string
          format: date-time
          description: Defaults to now

    # ========== REPORTS ==========
...
    AssignmentReport:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        dateRange:
          type: object
          properties:
            from:
              type: string
              format: date
            to:
              type: string
              format: date
        totalAssignments:
          type: integer
        totalReturns:
          type: integer
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentDetail'
...
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
        total:
          type: integer

    # Constitution Art. 4.2 mandates this error format
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code (e.g., VALIDATION_ERROR)
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string